TZ_news_map/requirements.txt
requests
psycopg2-binary
PySide6
